fix: Only clamp MFCC bounds when they exceed Nyquist, preserve 25 Hz band

+++ src/paradigma/pipelines/tremor_pipeline.py
@@ -337,10 +337,10 @@ def extract_spectral_domain_features(data: np.ndarray, config) -> pd.DataFrame:
     window_type = "hann"
 
     # Compute the power spectral density
-    segment_length_n = sampling_frequency * segment_length_psd_s
-    overlap_n = segment_length_n * overlap_fraction
+    segment_length_n = int(sampling_frequency * segment_length_psd_s)
+    overlap_n = int(segment_length_n * overlap_fraction)
     window = signal.get_window(window_type, segment_length_n, fftbins=False)
-    nfft = sampling_frequency / spectral_resolution
+    nfft = int(sampling_frequency / spectral_resolution)
 
     freqs, psd = signal.welch(
         x=data,
@@ -355,8 +355,8 @@ def extract_spectral_domain_features(data: np.ndarray, config) -> pd.DataFrame:
     )
 
     # Compute the spectrogram
-    segment_length_n = sampling_frequency * segment_length_spectrogram_s
-    overlap_n = segment_length_n * overlap_fraction
+    segment_length_n = int(sampling_frequency * segment_length_spectrogram_s)
+    overlap_n = int(segment_length_n * overlap_fraction)
     window = signal.get_window(window_type, segment_length_n)
 
     f, t, stft_result = signal.stft(
